extract vocabulary functions
