+++ onmt/data/init.lua
@@ -2,5 +2,6 @@ local data = {}
 
 data.Dataset = require('onmt.data.Dataset')
 data.Batch = require('onmt.data.Batch')
+data.Vocabulary = require('onmt.data.Vocabulary')
 
 return data
