+++ preprocess.lua
@@ -37,127 +37,8 @@ onmt.utils.Logger.declareOpts(cmd)
 
 local opt = cmd:parse(arg)
 
-local function hasFeatures(filename)
-  local reader = onmt.utils.FileReader.new(filename)
-  local _, _, numFeatures = onmt.utils.Features.extract(reader:next())
-  reader:close()
-  return numFeatures > 0
-end
-
 local function isValid(sent, maxSeqLength)
   return #sent > 0 and #sent <= maxSeqLength
-end
-
-local function makeVocabulary(filename, size, maxSeqLength)
-  local wordVocab = onmt.utils.Dict.new({onmt.Constants.PAD_WORD, onmt.Constants.UNK_WORD,
-                                         onmt.Constants.BOS_WORD, onmt.Constants.EOS_WORD})
-  local featuresVocabs = {}
-
-  local reader = onmt.utils.FileReader.new(filename)
-
-  while true do
-    local sent = reader:next()
-    if sent == nil then
-      break
-    end
-
-    if isValid(sent, maxSeqLength) then
-      local words, features, numFeatures = onmt.utils.Features.extract(sent)
-
-      if #featuresVocabs == 0 and numFeatures > 0 then
-        for j = 1, numFeatures do
-          featuresVocabs[j] = onmt.utils.Dict.new({onmt.Constants.PAD_WORD, onmt.Constants.UNK_WORD,
-                                                   onmt.Constants.BOS_WORD, onmt.Constants.EOS_WORD})
-        end
-      else
-        assert(#featuresVocabs == numFeatures,
-               'all sentences must have the same numbers of additional features')
-      end
-
-      for i = 1, #words do
-        wordVocab:add(words[i])
-
-        for j = 1, numFeatures do
-          featuresVocabs[j]:add(features[j][i])
-        end
-      end
-    end
-
-  end
-
-  reader:close()
-
-  local originalSize = wordVocab:size()
-  wordVocab = wordVocab:prune(size)
-  _G.logger:info('Created dictionary of size ' .. wordVocab:size() .. ' (pruned from ' .. originalSize .. ')')
-
-  return wordVocab, featuresVocabs
-end
-
-local function initVocabulary(name, dataFile, vocabFile, vocabSize, featuresVocabsFiles, maxSeqLength)
-  local wordVocab
-  local featuresVocabs = {}
-
-  if vocabFile:len() > 0 then
-    -- If given, load existing word dictionary.
-    _G.logger:info('Reading ' .. name .. ' vocabulary from \'' .. vocabFile .. '\'...')
-    wordVocab = onmt.utils.Dict.new()
-    wordVocab:loadFile(vocabFile)
-    _G.logger:info('Loaded ' .. wordVocab:size() .. ' ' .. name .. ' words')
-  end
-
-  if featuresVocabsFiles:len() > 0 then
-    -- If given, discover existing features dictionaries.
-    local j = 1
-
-    while true do
-      local file = featuresVocabsFiles .. '.' .. name .. '_feature_' .. j .. '.dict'
-
-      if not path.exists(file) then
-        break
-      end
-
-      _G.logger:info('Reading ' .. name .. ' feature ' .. j .. ' vocabulary from \'' .. file .. '\'...')
-      featuresVocabs[j] = onmt.utils.Dict.new()
-      featuresVocabs[j]:loadFile(file)
-      _G.logger:info('Loaded ' .. featuresVocabs[j]:size() .. ' labels')
-
-      j = j + 1
-    end
-  end
-
-  if wordVocab == nil or (#featuresVocabs == 0 and hasFeatures(dataFile)) then
-    -- If a dictionary is still missing, generate it.
-    _G.logger:info('Building ' .. name  .. ' vocabulary...')
-    local genWordVocab, genFeaturesVocabs = makeVocabulary(dataFile, vocabSize, maxSeqLength)
-
-    if wordVocab == nil then
-      wordVocab = genWordVocab
-    end
-    if #featuresVocabs == 0 then
-      featuresVocabs = genFeaturesVocabs
-    end
-  end
-
-  _G.logger:info('')
-
-  return {
-    words = wordVocab,
-    features = featuresVocabs
-  }
-end
-
-local function saveVocabulary(name, vocab, file)
-  _G.logger:info('Saving ' .. name .. ' vocabulary to \'' .. file .. '\'...')
-  vocab:writeFile(file)
-end
-
-local function saveFeaturesVocabularies(name, vocabs, prefix)
-  for j = 1, #vocabs do
-    local file = prefix .. '.' .. name .. '_feature_' .. j .. '.dict'
-    _G.logger:info('Saving ' .. name .. ' feature ' .. j .. ' vocabulary to \'' .. file .. '\'...')
-    vocabs[j]:writeFile(file)
-  end
 end
 
 local function vecToTensor(vec)
@@ -279,13 +160,15 @@ local function main()
 
   _G.logger = onmt.utils.Logger.new(opt.log_file, opt.disable_logs, opt.log_level)
 
+  local Vocabulary = onmt.data.Vocabulary
+
   local data = {}
 
   data.dicts = {}
-  data.dicts.src = initVocabulary('source', opt.train_src, opt.src_vocab, opt.src_vocab_size,
-                                  opt.features_vocabs_prefix, opt.src_seq_length)
-  data.dicts.tgt = initVocabulary('target', opt.train_tgt, opt.tgt_vocab, opt.tgt_vocab_size,
-                                  opt.features_vocabs_prefix, opt.tgt_seq_length)
+  data.dicts.src = Vocabulary.init('source', opt.train_src, opt.src_vocab, opt.src_vocab_size,
+                                   opt.features_vocabs_prefix, function(s) return isValid(s, opt.src_seq_length) end)
+  data.dicts.tgt = Vocabulary.init('target', opt.train_tgt, opt.tgt_vocab, opt.tgt_vocab_size,
+                                   opt.features_vocabs_prefix, function(s) return isValid(s, opt.tgt_seq_length) end)
 
   _G.logger:info('Preparing training data...')
   data.train = {}
@@ -300,16 +183,16 @@ local function main()
   _G.logger:info('')
 
   if opt.src_vocab:len() == 0 then
-    saveVocabulary('source', data.dicts.src.words, opt.save_data .. '.src.dict')
+    Vocabulary.save('source', data.dicts.src.words, opt.save_data .. '.src.dict')
   end
 
   if opt.tgt_vocab:len() == 0 then
-    saveVocabulary('target', data.dicts.tgt.words, opt.save_data .. '.tgt.dict')
+    Vocabulary.save('target', data.dicts.tgt.words, opt.save_data .. '.tgt.dict')
   end
 
   if opt.features_vocabs_prefix:len() == 0 then
-    saveFeaturesVocabularies('source', data.dicts.src.features, opt.save_data)
-    saveFeaturesVocabularies('target', data.dicts.tgt.features, opt.save_data)
+    Vocabulary.saveFeatures('source', data.dicts.src.features, opt.save_data)
+    Vocabulary.saveFeatures('target', data.dicts.tgt.features, opt.save_data)
   end
 
   _G.logger:info('Saving data to \'' .. opt.save_data .. '-train.t7\'...')
